pyproject.toml
[project]
name = "hotel-booking-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "black>=25.1.0",
    "celery>=5.5.3",
    "fastapi-cache2>=0.2.2",
    "fastapi[standard]>=0.116.1",
    "greenlet>=3.2.4",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pyright>=1.1.405",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-dotenv>=0.5.2",
    "redis>=6.4.0",
    "ruff>=0.13.2",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.43",
]
[tool.ruff]
line-length = 100
